Hack to fix issues with pip installation/homebrew.
This is documented in #16 but basically just trying to get a build going
without seeing any errors. Not sure what a long term fix is
here. Possibly the Boxen environment needs to be passed into the type
providers?

Another possible solution is just using exec as I believe it was solved
placing this in the PATH for site.pp in our-boxen.